requirements.txt
aiogram==2.5
openai==0.28.1
python-dotenv
